change tests and paths
